+++ TransferModels/DemoDownloadReport.cs
@@ -2,7 +2,10 @@
 
 namespace RabbitCommunicationLib.TransferModels
 {
-    public class DemoObtainReport : TransferModel, IBlobUrl, IMatchId
+    /// <summary>
+    /// Report regarding the download of a demo file.
+    /// </summary>
+    public class DemoDownloadReport : TransferModel, IBlobUrl, IMatchId
     {
         public long MatchId { get; set; }