+++ TransferModels/DemoRemovalInstruction.cs
@@ -6,6 +6,9 @@ using RabbitCommunicationLib.TransferModels.Interfaces;
 
 namespace RabbitCommunicationLib.TransferModels
 {
+    /// <summary>
+    /// Instructions for MatchWriter to remove a MatchDataSet from the MatchDb.
+    /// </summary>
     public class DemoRemovalInstruction : TransferModel, IMatchId
     {
         public long MatchId { get ; set; }