+++ TransferModels/MatchDatabaseInsertionInstruction.cs
@@ -5,7 +5,10 @@ using System.Text;
 
 namespace RabbitCommunicationLib.TransferModels
 {
-    public class RedisLocalizationInstruction : TransferModel, IMatchId
+    /// <summary>
+    /// Instructions for MatchWriter to upload of a MatchDataSet into the MatchDb.
+    /// </summary>
+    public class MatchDatabaseInsertionInstruction : TransferModel, IMatchId
     {
         /// <inheritdoc/>
         public long MatchId { get; set; }