+++ TransferModels/TaskCompletedReport.cs
@@ -12,9 +12,10 @@ namespace RabbitCommunicationLib.TransferModels
         /// </summary>
         public bool Success { get; set; }
 
-        /// <summary>
-        /// The version of the service that completed the task.
-        /// </summary>
-        public string Version { get; set; }
+        public TaskCompletedReport(long matchId)
+        {
+            MatchId = matchId;
+            Success = false;
+        }
     }
 }