+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2020-05-26
+##Added
+- SituationExtractionReport
+- MatchDatabaseInsertionReport (MatchWriter was previously using TaskCompletedReport)
+- Constructor for TaskCompletedReport
+- Added enum values for SituationOperator to DemoAnalysisBlock
+
+### Changed
+- Make DemoAnalysisBlock props nullable and default to null
+- Rename DemoObtainReport to DemoDownloadReport
+- Rename ManualDownloadReport to ManualDownloadInsertInstruction
+- Rename RedisLocalizationInstruction to MatchDatabaseInsertionInstruction
+
+### Removed
+- Property DemoAnalyzeReport.FramesPerSecond
+
 ## [0.9.0] - 2020-05-25
 ### Changed
 - DemoAnalyzeReport, added Block field to indicate reason for Analysis blocking.